Now starts FormalAsync. Fixed a couple of bugs.
Not the image generated still has a fixed size. This needs to be looked
at.

+++ app/src/main/java/com/johnston/lmhapp/MealMenus/NotificationsService.java
@@ -108,7 +108,7 @@ public class NotificationsService extends BroadcastReceiver {
     public void part3(File file,Context context){
         try {
             SharedPreferences refreshTimePreference = context.getSharedPreferences("RefreshTime",0);
-            long  refreshTime = refreshTimePreference.getInt("refreshTime",2);
+            long  refreshTime = refreshTimePreference.getLong("refreshTime",2);
             WidgetBroadcastReceiver mealMenu = new WidgetBroadcastReceiver();
             BufferedReader br = new BufferedReader(new FileReader(file));
 //            Check the date.
